requirements.txt
requests==2.31.0
fake-useragent==1.4.0
beautifulsoup4==4.12.2
pandas==2.2.1
sqlalchemy==2.0.29
lxml==5.2.2
html5lib==1.1
python-dateutil==2.8.2
pydantic==2.5.2
flask==3.0.2
schedule==1.2.1
python-dotenv==0.19.0